+++ GANDLF/compute/training_loop.py
@@ -16,11 +16,14 @@ from GANDLF.utils import (
     get_date_time,
     send_model_to_device,
     populate_channel_keys_in_params,
-    get_class_imbalance_weights,
     save_model,
     load_model,
     version_check,
     write_training_patches,
+)
+from GANDLF.utils.tensor import (
+    get_class_imbalance_weights_segmentation,
+    get_class_imbalance_weights_classification,
 )
 from GANDLF.logger import Logger
 from .step import step
@@ -219,13 +222,16 @@ def training_loop(
     # Fetch the model according to params mentioned in the configuration file
     model = global_models_dict[params["model"]["architecture"]](parameters=params)
 
+    training_data_copy = training_data.copy()
+    validation_data_copy = validation_data.copy()
+
     # Set up the dataloaders
     training_data_for_torch = ImagesFromDataFrame(
-        training_data, params, train=True, loader_type="train"
+        training_data_copy, params, train=True, loader_type="train"
     )
 
     validation_data_for_torch = ImagesFromDataFrame(
-        validation_data, params, train=False, loader_type="validation"
+        validation_data_copy, params, train=False, loader_type="validation"
     )
 
     testingDataDefined = True
@@ -312,25 +318,35 @@ def training_loop(
 
     # Calculate the weights here
     if params["weighted_loss"]:
-        print("Calculating weights for loss")
-        # Set up the dataloader for penalty calculation
-        penalty_data = ImagesFromDataFrame(
-            training_data,
-            parameters=params,
-            train=False,
-            loader_type="penalty",
-        )
-        penalty_loader = DataLoader(
-            penalty_data,
-            batch_size=1,
-            shuffle=True,
-            pin_memory=False,
-        )
+        print("Calculating weights")
+        # if params["weighted_loss"][weights] is None # You can get weights from the user here, might need some playing with class_list to do later
+        if params["problem_type"] == "classification":
+            (
+                params["weights"],
+                params["class_weights"],
+            ) = get_class_imbalance_weights_classification(training_data, params)
+        elif params["problem_type"] == "segmentation":
+            print("Calculating weights for loss")
+            # Set up the dataloader for penalty calculation
+            penalty_data = ImagesFromDataFrame(
+                training_data,
+                parameters=params,
+                train=False,
+                loader_type="penalty",
+            )
 
-        params["weights"], params["class_weights"] = get_class_imbalance_weights(
-            penalty_loader, params
-        )
-        del penalty_data, penalty_loader
+            penalty_loader = DataLoader(
+                penalty_data,
+                batch_size=1,
+                shuffle=True,
+                pin_memory=False,
+            )
+
+            (
+                params["weights"],
+                params["class_weights"],
+            ) = get_class_imbalance_weights_segmentation(penalty_loader, params)
+            del penalty_data, penalty_loader
     else:
         params["weights"], params["class_weights"] = None, None
 
@@ -358,7 +374,6 @@ def training_loop(
 
     # if previous model file is present, load it up
     if os.path.exists(best_model_path):
-        print("Previous model found. Loading it up.")
         try:
             main_dict = load_model(best_model_path)
             version_check(params["version"], version_to_check=main_dict["version"])
@@ -366,9 +381,9 @@ def training_loop(
             start_epoch = main_dict["epoch"]
             optimizer.load_state_dict(main_dict["optimizer_state_dict"])
             best_loss = main_dict["loss"]
-            print("Previous model loaded successfully.")
-        except IOError:
-            raise IOError("Previous model could not be loaded, error: ")
+            print("Previous model successfully loaded.")
+        except RuntimeWarning:
+            RuntimeWarning("Previous model could not be loaded, initializing model")
 
     print("Using device:", device, flush=True)
 
@@ -460,6 +475,8 @@ def training_loop(
             best_loss = epoch_valid_loss
             best_train_idx = epoch
             patience = 0
+
+            model.eval()
             save_model(
                 {
                     "epoch": best_train_idx,
@@ -467,8 +484,12 @@ def training_loop(
                     "optimizer_state_dict": optimizer.state_dict(),
                     "loss": best_loss,
                 },
+                model,
+                params,
                 best_model_path,
+                onnx_export=False,
             )
+            model.train()
             first_model_saved = True
 
         print("Current Best epoch: ", best_train_idx)
@@ -490,6 +511,42 @@ def training_loop(
         " mins",
         flush=True,
     )
+
+    # once the training is done, optimize the best model
+    if os.path.exists(best_model_path):
+
+        onnx_export = True
+        if params["model"]["architecture"] in ["sdnet", "brain_age"]:
+            onnx_export = False
+        elif (
+            "onnx_export" in params["model"] and params["model"]["onnx_export"] == False
+        ):
+            onnx_export = False
+
+        if onnx_export:
+            print("Optimizing best model.")
+
+            try:
+                main_dict = load_model(best_model_path)
+                version_check(params["version"], version_to_check=main_dict["version"])
+                model.load_state_dict(main_dict["model_state_dict"])
+                best_epoch = main_dict["epoch"]
+                optimizer.load_state_dict(main_dict["optimizer_state_dict"])
+                best_loss = main_dict["loss"]
+                save_model(
+                    {
+                        "epoch": best_epoch,
+                        "model_state_dict": model.state_dict(),
+                        "optimizer_state_dict": optimizer.state_dict(),
+                        "loss": best_loss,
+                    },
+                    model,
+                    params,
+                    best_model_path,
+                    onnx_export,
+                )
+            except Exception as e:
+                print("Best model could not be loaded, error: ", e)
 
 
 if __name__ == "__main__":