+++ testing/test_full.py
@@ -2,6 +2,7 @@ from pathlib import Path
 import requests, zipfile, io, os, csv, random, copy, shutil, sys, yaml, torch, pytest
 import SimpleITK as sitk
 import numpy as np
+import pandas as pd
 
 from pydicom.data import get_testdata_file
 
@@ -16,7 +17,7 @@ from GANDLF.cli import main_run, preprocess_and_save, patch_extraction
 from GANDLF.schedulers import global_schedulers_dict
 from GANDLF.optimizers import global_optimizer_dict
 from GANDLF.models import global_models_dict
-from GANDLF.post_process import torch_morphological, fill_holes
+from GANDLF.data.post_process import torch_morphological, fill_holes, get_mapped_label
 from GANDLF.anonymize import run_anonymizer
 
 device = "cpu"
@@ -24,7 +25,9 @@ device = "cpu"
 # pre-defined segmentation model types for testing
 all_models_segmentation = [
     "lightunet",
+    "lightunet_multilayer",
     "unet",
+    "unet_multilayer",
     "deep_resunet",
     "fcn",
     "uinc",
@@ -38,8 +41,23 @@ all_models_regression = [
     "resnet50",
     "efficientnetb0",
 ]
+# pre-defined regression/classification model types for testing
+all_models_classification = [
+    "imagenet_vgg11",
+    "imagenet_vgg11_bn",
+    "imagenet_vgg13",
+    "imagenet_vgg13_bn",
+    "imagenet_vgg16",
+    "imagenet_vgg16_bn",
+    "imagenet_vgg19",
+    "imagenet_vgg19_bn",
+    "resnet18",
+]
+
 all_clip_modes = ["norm", "value", "agc"]
 all_norm_type = ["batch", "instance"]
+
+all_model_type = ["torch", "openvino"]
 
 patch_size = {"2D": [128, 128, 1], "3D": [32, 32, 32]}
 
@@ -162,11 +180,13 @@ def test_train_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
     parameters["model"]["amp"] = True
     parameters["model"]["num_channels"] = 3
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # read and initialize parameters for specific data dimension
     for model in all_models_segmentation:
@@ -202,6 +222,7 @@ def test_train_segmentation_sdnet_rad_2d(device):
     parameters["model"]["class_list"] = [0, 255]
     parameters["model"]["num_channels"] = 1
     parameters["model"]["architecture"] = "sdnet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     sanitize_outputDir()
     TrainingManager(
@@ -226,12 +247,14 @@ def test_train_segmentation_rad_3d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["model"]["dimension"] = 3
     parameters["model"]["class_list"] = [0, 1]
     parameters["model"]["amp"] = True
     parameters["in_memory"] = True
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_segmentation:
@@ -252,10 +275,12 @@ def test_train_segmentation_rad_3d(device):
 
 
 def test_train_regression_rad_2d(device):
+    print("Starting 2D Rad regression tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["amp"] = False
@@ -266,6 +291,7 @@ def test_train_regression_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
     parameters["scaling_factor"] = 1
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
@@ -285,8 +311,10 @@ def test_train_regression_rad_2d(device):
     print("passed")
 
 
-def test_train_brainage_rad_2d(device):
+def test_train_regression_rad_2d_imagenet(device):
+    print("Starting 2D Rad regression tests for imagenet models")
     # read and initialize parameters for specific data dimension
+    print("Starting 2D Rad regression tests for imagenet models")
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
@@ -300,7 +328,44 @@ def test_train_brainage_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
     parameters["scaling_factor"] = 1
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    # loop through selected models and train for single epoch
+    for model in all_models_classification:
+        parameters["model"]["architecture"] = model
+        parameters["nested_training"]["testing"] = 1
+        parameters["nested_training"]["validation"] = -5
+        sanitize_outputDir()
+        TrainingManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+            resume=False,
+            reset=True,
+        )
+
+    print("passed")
+
+
+def test_train_brainage_rad_2d(device):
+    print("Starting brain age tests")
+    # read and initialize parameters for specific data dimension
+    parameters = parseConfig(
+        testingDir + "/config_regression.yaml", version_check_flag=False
+    )
+    parameters["modality"] = "rad"
+    parameters["patch_size"] = patch_size["2D"]
+    parameters["model"]["dimension"] = 2
+    parameters["model"]["amp"] = False
+    # read and parse csv
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_regression.csv"
+    )
+    parameters["model"]["num_channels"] = 3
+    parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
+    parameters["scaling_factor"] = 1
     parameters["model"]["architecture"] = "brain_age"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     sanitize_outputDir()
     TrainingManager(
@@ -316,10 +381,12 @@ def test_train_brainage_rad_2d(device):
 
 
 def test_train_regression_rad_3d(device):
+    print("Starting 3D Rad regression tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["model"]["dimension"] = 3
     # read and parse csv
@@ -328,6 +395,7 @@ def test_train_regression_rad_3d(device):
     )
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
@@ -352,6 +420,7 @@ def test_train_regression_rad_3d(device):
 
 
 def test_train_classification_rad_2d(device):
+    print("Starting 2D Rad classification tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -365,6 +434,7 @@ def test_train_classification_rad_2d(device):
         inputDir + "/train_2d_rad_classification.csv"
     )
     parameters["model"]["num_channels"] = 3
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
@@ -385,6 +455,7 @@ def test_train_classification_rad_2d(device):
 
 
 def test_train_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -398,6 +469,7 @@ def test_train_classification_rad_3d(device):
     )
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["onnx_export"] = False
     # loop through selected models and train for single epoch
     for model in all_models_regression:
         if "efficientnet" in model:
@@ -420,7 +492,8 @@ def test_train_classification_rad_3d(device):
     print("passed")
 
 
-def test_inference_classification_rad_3d(device):
+def test_train_resume_inference_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests for resume and reset")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -437,6 +510,7 @@ def test_inference_classification_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
+    parameters["model"]["onnx_export"] = False
     sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
@@ -446,6 +520,7 @@ def test_inference_classification_rad_3d(device):
         resume=False,
         reset=True,
     )
+
     ## testing resume with parameter updates
     parameters["num_epochs"] = 2
     parameters["nested_training"]["testing"] = -5
@@ -460,7 +535,9 @@ def test_inference_classification_rad_3d(device):
     )
 
     ## testing resume without parameter updates
-    parameters["num_epochs"] = 3
+    parameters["num_epochs"] = 1
+    parameters["nested_training"]["testing"] = -5
+    parameters["nested_training"]["validation"] = -5
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -470,8 +547,7 @@ def test_inference_classification_rad_3d(device):
         reset=False,
     )
 
-    ## testing inference
-    parameters["output_dir"] = outputDir
+    parameters["output_dir"] = outputDir  # this is in inference mode
     InferenceManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -482,7 +558,93 @@ def test_inference_classification_rad_3d(device):
     print("passed")
 
 
+def test_inference_optimize_classification_rad_3d(device):
+    print("Starting 3D Rad segmentation tests for optimization")
+    # read and initialize parameters for specific data dimension
+    parameters = parseConfig(
+        testingDir + "/config_classification.yaml", version_check_flag=False
+    )
+    parameters["modality"] = "rad"
+    parameters["patch_size"] = patch_size["3D"]
+    parameters["model"]["dimension"] = 3
+    # read and parse csv
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_3d_rad_classification.csv"
+    )
+    parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["architecture"] = all_models_regression[0]
+    parameters["model"]["onnx_export"] = True
+    sanitize_outputDir()
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=True,
+    )
+
+    ## testing inference
+    for model_type in all_model_type:
+        parameters["model"]["type"] = model_type
+        parameters["output_dir"] = outputDir  # this is in inference mode
+        InferenceManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+        )
+
+    print("passed")
+
+
+def test_inference_optimize_segmentation_rad_2d(device):
+    print("Starting 2D Rad segmentation tests for optimization")
+    # read and parse csv
+    parameters = parseConfig(
+        testingDir + "/config_segmentation.yaml", version_check_flag=False
+    )
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    parameters["patch_size"] = patch_size["2D"]
+    parameters["modality"] = "rad"
+    parameters["model"]["dimension"] = 2
+    parameters["model"]["class_list"] = [0, 255]
+    parameters["model"]["amp"] = True
+    parameters["save_output"] = True
+    parameters["model"]["num_channels"] = 3
+    parameters["metrics"] = ["dice"]
+    parameters["model"]["architecture"] = "resunet"
+    parameters["model"]["onnx_export"] = True
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    sanitize_outputDir()
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=True,
+    )
+
+    ## testing inference
+    for model_type in all_model_type:
+        parameters["model"]["type"] = model_type
+        parameters["output_dir"] = outputDir  # this is in inference mode
+        InferenceManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+        )
+
+    print("passed")
+
+
 def test_inference_classification_with_logits_single_fold_rad_3d(device):
+    print("Starting 3D Rad classification tests for single fold logits inference")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -501,6 +663,7 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
+    parameters["model"]["onnx_export"] = False
     sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
@@ -510,7 +673,28 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
         resume=False,
         reset=True,
     )
+    ## this is to test if inference can run without having ground truth column
+    training_data.drop("ValueToPredict", axis=1, inplace=True)
+    training_data.drop("Label", axis=1, inplace=True)
+    temp_infer_csv = os.path.join(outputDir, "temp_infer_csv.csv")
+    training_data.to_csv(temp_infer_csv, index=False)
+    # read and parse csv
+    parameters = parseConfig(
+        testingDir + "/config_classification.yaml", version_check_flag=False
+    )
+    training_data, parameters["headers"] = parseTrainingCSV(temp_infer_csv)
     parameters["output_dir"] = outputDir  # this is in inference mode
+    parameters["output_dir"] = outputDir  # this is in inference mode
+    parameters["modality"] = "rad"
+    parameters["patch_size"] = patch_size["3D"]
+    parameters["model"]["dimension"] = 3
+    parameters["model"]["final_layer"] = "logits"
+    parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    # loop through selected models and train for single epoch
+    model = all_models_regression[0]
+    parameters["model"]["architecture"] = model
+    parameters["model"]["onnx_export"] = False
     InferenceManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -522,6 +706,7 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
 
 
 def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
+    print("Starting 3D Rad classification tests for multi-fold logits inference")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -532,7 +717,7 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
     parameters["model"]["final_layer"] = "logits"
     # necessary for n-fold cross-validation inference
     parameters["nested_training"]["validation"] = 2
-
+    parameters["model"]["onnx_export"] = False
     # read and parse csv
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_classification.csv"
@@ -563,6 +748,7 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
 
 
 def test_scheduler_classification_rad_2d(device):
+    print("Starting 2D Rad segmentation tests for scheduler")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -578,6 +764,7 @@ def test_scheduler_classification_rad_2d(device):
     parameters["model"]["architecture"] = "densenet121"
     parameters["model"]["norm_type"] = "instance"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["onnx_export"] = False
     # loop through selected models and train for single epoch
     for scheduler in global_schedulers_dict:
         parameters["scheduler"] = {}
@@ -600,6 +787,7 @@ def test_scheduler_classification_rad_2d(device):
 
 
 def test_optimizer_classification_rad_2d(device):
+    print("Starting 2D Rad classification tests for optimizer")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -614,6 +802,7 @@ def test_optimizer_classification_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "densenet121"
     parameters["model"]["norm_type"] = "none"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for optimizer in global_optimizer_dict:
@@ -637,6 +826,7 @@ def test_optimizer_classification_rad_2d(device):
 
 
 def test_clip_train_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests for clipping")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -651,6 +841,7 @@ def test_clip_train_classification_rad_3d(device):
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["architecture"] = "vgg16"
     parameters["model"]["norm_type"] = "None"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for clip_mode in all_clip_modes:
@@ -670,8 +861,8 @@ def test_clip_train_classification_rad_3d(device):
 
 
 def test_normtype_train_segmentation_rad_3d(device):
-    # read and initialize parameters for specific data dimension
     print("Starting 3D Rad segmentation tests for normtype")
+    # read and initialize parameters for specific data dimension
     # read and parse csv
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
@@ -687,6 +878,7 @@ def test_normtype_train_segmentation_rad_3d(device):
     parameters["save_output"] = True
     parameters["data_postprocessing"] = {"fill_holes"}
     parameters["in_memory"] = True
+    parameters["model"]["onnx_export"] = False
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
@@ -720,6 +912,7 @@ def test_metrics_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -728,6 +921,7 @@ def test_metrics_segmentation_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["metrics"] = ["dice", "hausdorff", "hausdorff95"]
     parameters["model"]["architecture"] = "resunet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     sanitize_outputDir()
     TrainingManager(
@@ -751,6 +945,7 @@ def test_metrics_regression_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_regression.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -758,6 +953,7 @@ def test_metrics_regression_rad_2d(device):
     parameters["model"]["amp"] = False
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "vgg11"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     sanitize_outputDir()
     TrainingManager(
@@ -781,6 +977,7 @@ def test_losses_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -789,6 +986,7 @@ def test_losses_segmentation_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "resunet"
     parameters["metrics"] = ["dice"]
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for loss_type in ["dc", "dc_log", "dcce", "dcce_logits", "tversky"]:
@@ -859,7 +1057,7 @@ def test_config_read():
 
     os.remove(file_config_temp)
 
-    # ensure resize_image is triggered
+    # ensure resize_patch is triggered
     parameters["data_preprocessing"].pop("resize_image")
     parameters["data_preprocessing"]["resize_patch"] = [64, 64]
 
@@ -910,6 +1108,7 @@ def test_cli_function_preprocess():
     file_data = os.path.join(inputDir, "train_2d_rad_segmentation.csv")
 
     parameters = parseConfig(file_config)
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = "[0, 255||125]"
@@ -948,6 +1147,7 @@ def test_cli_function_mainrun(device):
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
 
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["num_epochs"] = 1
     parameters["nested_training"]["testing"] = 1
@@ -1000,14 +1200,19 @@ def test_dataloader_construction_train_segmentation_3d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["save_training"] = True
+    parameters["save_output"] = True
     parameters["model"]["dimension"] = 3
     parameters["model"]["class_list"] = [0, 1]
     parameters["model"]["amp"] = True
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["architecture"] = "unet"
     parameters["weighted_loss"] = False
+    parameters["model"]["onnx_export"] = False
+    parameters["data_postprocessing"]["mapping"] = {0: 0, 1: 1}
+    parameters["data_postprocessing"]["fill_holes"] = True
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     sanitize_outputDir()
@@ -1145,6 +1350,7 @@ def test_checkpointing_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["num_epochs"] = 1
     parameters["nested_training"]["testing"] = 1
@@ -1161,6 +1367,7 @@ def test_checkpointing_segmentation_rad_2d(device):
         "hd100_per_label",
     ]
     parameters["model"]["architecture"] = "unet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     sanitize_outputDir()
     TrainingManager(
@@ -1191,7 +1398,7 @@ def test_model_patch_divisibility():
     parameters = parseConfig(
         testingDir + "/config_segmentation.yaml", version_check_flag=False
     )
-    training_data, parameters["headers"] = parseTrainingCSV(
+    _, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
     parameters["model"]["architecture"] = "unet"
@@ -1206,14 +1413,14 @@ def test_model_patch_divisibility():
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
 
     # this assertion should fail
-    with pytest.raises(Exception) as e_info:
+    with pytest.raises(BaseException) as e_info:
         global_models_dict[parameters["model"]["architecture"]](parameters=parameters)
 
     parameters["model"]["architecture"] = "uinc"
     parameters["model"]["base_filters"] = 11
 
     # this assertion should fail
-    with pytest.raises(Exception) as e_info:
+    with pytest.raises(BaseException) as e_info:
         global_models_dict[parameters["model"]["architecture"]](parameters=parameters)
 
     print("passed")
@@ -1254,6 +1461,15 @@ def test_one_hot_logic():
     )
     comparison = combined_array == (img_tensor_oh_rev_array == 1)
     assert comparison.all(), "Arrays at the combined foreground are not equal"
+
+    parameters = {"data_postprocessing": {"mapping": {0: 0, 1: 1, 2: 5}}}
+    mapped_output = get_mapped_label(
+        torch.from_numpy(img_tensor_oh_rev_array), parameters
+    )
+
+    for key, value in parameters["data_postprocessing"]["mapping"].items():
+        comparison = (img_tensor_oh_rev_array == key) == (mapped_output == value)
+        assert comparison.all(), "Arrays at {}:{} are not equal".format(key, value)
 
     print("passed")
 
@@ -1304,7 +1520,7 @@ def test_anonymizer():
 
 
 def test_train_inference_segmentation_histology_2d(device):
-    print("Starting histology train/inference tests")
+    print("Starting histology train/inference segmentation tests")
     # overwrite previous results
     sanitize_outputDir()
     output_dir_patches = os.path.join(outputDir, "histo_patches")
@@ -1351,6 +1567,82 @@ def test_train_inference_segmentation_histology_2d(device):
     parameters["nested_training"]["testing"] = 1
     parameters["nested_training"]["validation"] = -2
     parameters["metrics"] = ["dice"]
+    parameters["model"]["onnx_export"] = True
+    modelDir = os.path.join(outputDir, "modelDir")
+    Path(modelDir).mkdir(parents=True, exist_ok=True)
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=modelDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=True,
+    )
+    inference_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_histo_segmentation.csv", train=False
+    )
+    inference_data.drop(index=inference_data.index[-1], axis=0, inplace=True)
+    InferenceManager(
+        dataframe=inference_data,
+        outputDir=modelDir,
+        parameters=parameters,
+        device=device,
+    )
+
+    print("passed")
+
+
+def test_train_inference_classification_histology_2d(device):
+    print("Starting histology train/inference classification tests")
+    # overwrite previous results
+    sanitize_outputDir()
+    output_dir_patches = os.path.join(outputDir, "histo_patches")
+    if os.path.isdir(output_dir_patches):
+        shutil.rmtree(output_dir_patches)
+    Path(output_dir_patches).mkdir(parents=True, exist_ok=True)
+    output_dir_patches_output = os.path.join(output_dir_patches, "histo_patches_output")
+    Path(output_dir_patches_output).mkdir(parents=True, exist_ok=True)
+    file_config_temp = os.path.join(
+        output_dir_patches, "config_patch-extraction_temp.yaml"
+    )
+    # if found in previous run, discard.
+    if os.path.exists(file_config_temp):
+        os.remove(file_config_temp)
+
+    parameters_patch = {}
+    # extracting minimal number of patches to ensure that the test does not take too long
+    parameters_patch["num_patches"] = 3
+    parameters_patch["patch_size"] = [128, 128]
+
+    with open(file_config_temp, "w") as file:
+        yaml.dump(parameters_patch, file)
+
+    patch_extraction(
+        inputDir + "/train_2d_histo_classification.csv",
+        output_dir_patches_output,
+        file_config_temp,
+    )
+
+    file_for_Training = os.path.join(output_dir_patches_output, "opm_train.csv")
+    temp_df = pd.read_csv(file_for_Training)
+    temp_df.drop("Label", axis=1, inplace=True)
+    temp_df["valuetopredict"] = np.random.randint(2, size=6)
+    temp_df.to_csv(file_for_Training, index=False)
+    # read and parse csv
+    parameters = parseConfig(
+        testingDir + "/config_classification.yaml", version_check_flag=False
+    )
+    parameters["modality"] = "histo"
+    parameters["patch_size"] = patch_size["2D"]
+    parameters["model"]["dimension"] = 2
+    # read and parse csv
+    training_data, parameters["headers"] = parseTrainingCSV(file_for_Training)
+    parameters["model"]["num_channels"] = 3
+    parameters["model"]["architecture"] = "densenet121"
+    parameters["model"]["norm_type"] = "none"
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["nested_training"]["testing"] = 1
+    parameters["nested_training"]["validation"] = -2
     modelDir = os.path.join(outputDir, "modelDir")
     if os.path.isdir(modelDir):
         shutil.rmtree(modelDir)
@@ -1365,13 +1657,68 @@ def test_train_inference_segmentation_histology_2d(device):
     )
     parameters["output_dir"] = modelDir  # this is in inference mode
     inference_data, parameters["headers"] = parseTrainingCSV(
-        inputDir + "/train_2d_histo_segmentation.csv", train=False
+        inputDir + "/train_2d_histo_classification.csv", train=False
     )
-    InferenceManager(
-        dataframe=inference_data,
-        outputDir=modelDir,
-        parameters=parameters,
-        device=device,
+    for model_type in all_model_type:
+        parameters["nested_training"]["testing"] = 1
+        parameters["nested_training"]["validation"] = -2
+        parameters["output_dir"] = modelDir  # this is in inference mode
+        inference_data, parameters["headers"] = parseTrainingCSV(
+            inputDir + "/train_2d_histo_segmentation.csv", train=False
+        )
+        parameters["model"]["type"] = model_type
+        InferenceManager(
+            dataframe=inference_data,
+            outputDir=modelDir,
+            parameters=parameters,
+            device=device,
+        )
+
+    print("passed")
+
+
+def test_unet_layerchange_2d(device):
+    # test case to up code coverage --> test decreasing allowed layers for unet
+    print("Starting 2D Rad segmentation tests for normtype")
+    # read and parse csv
+    # read and initialize parameters for specific data dimension
+    parameters = parseConfig(
+        testingDir + "/config_segmentation.yaml", version_check_flag=False
     )
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    for model in ["unet_multilayer", "lightunet_multilayer"]:
+        parameters["model"]["architecture"] = model
+        parameters["patch_size"] = [4, 4, 1]
+        parameters["model"]["dimension"] = 2
+
+        # this assertion should fail
+        with pytest.raises(BaseException) as e_info:
+            global_models_dict[parameters["model"]["architecture"]](
+                parameters=parameters
+            )
+
+        parameters["patch_size"] = patch_size["2D"]
+        parameters["model"]["depth"] = 7
+        parameters["model"]["class_list"] = [0, 255]
+        parameters["model"]["amp"] = True
+        parameters["model"]["num_channels"] = 3
+        parameters = populate_header_in_parameters(parameters, parameters["headers"])
+        # loop through selected models and train for single epoch
+        parameters["model"]["norm_type"] = "batch"
+        parameters["nested_training"]["testing"] = -5
+        parameters["nested_training"]["validation"] = -5
+        if os.path.isdir(outputDir):
+            shutil.rmtree(outputDir)  # overwrite previous results
+        sanitize_outputDir()
+        TrainingManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+            resume=False,
+            reset=True,
+        )
 
     print("passed")